+++ cli/src/cli_util.rs
@@ -129,6 +129,7 @@ use jj_lib::str_util::StringExpression;
 use jj_lib::str_util::StringMatcher;
 use jj_lib::str_util::StringPattern;
 use jj_lib::transaction::Transaction;
+use jj_lib::transaction::TransactionCommitError;
 use jj_lib::working_copy;
 use jj_lib::working_copy::CheckoutStats;
 use jj_lib::working_copy::LockedWorkingCopy;
@@ -415,6 +416,23 @@ impl CommandHelper {
             template_builder::parse(language, &mut diagnostics, template_text, &aliases)?;
         print_parse_diagnostics(ui, "In template expression", &diagnostics)?;
         Ok(template)
+    }
+
+    pub fn should_commit_transaction(&self) -> bool {
+        !self.global_args().no_integrate_operation
+    }
+
+    async fn maybe_commit_transaction(
+        &self,
+        tx: Transaction,
+        description: impl Into<String>,
+    ) -> Result<Arc<ReadonlyRepo>, TransactionCommitError> {
+        let unpublished_op = tx.write(description).await?;
+        if self.should_commit_transaction() {
+            unpublished_op.publish().await
+        } else {
+            Ok(unpublished_op.leave_unpublished())
+        }
     }
 
     pub fn workspace_loader(&self) -> Result<&dyn WorkspaceLoader, CommandError> {
@@ -1088,6 +1106,7 @@ pub struct WorkspaceCommandHelper {
     // TODO: Parsed template can be cached if it doesn't capture 'repo lifetime
     commit_summary_template_text: String,
     op_summary_template_text: String,
+    may_snapshot_working_copy: bool,
     may_update_working_copy: bool,
     working_copy_shared_with_git: bool,
 }
@@ -1125,8 +1144,10 @@ impl WorkspaceCommandHelper {
         let settings = workspace.settings();
         let commit_summary_template_text = settings.get_string("templates.commit_summary")?;
         let op_summary_template_text = settings.get_string("templates.op_summary")?;
-        let may_update_working_copy =
+        let may_snapshot_working_copy =
             loaded_at_head && !env.command.global_args().ignore_working_copy;
+        let may_update_working_copy =
+            may_snapshot_working_copy && env.command.should_commit_transaction();
         let working_copy_shared_with_git =
             crate::git_util::is_colocated_git_workspace(&workspace, &repo);
 
@@ -1136,6 +1157,7 @@ impl WorkspaceCommandHelper {
             env,
             commit_summary_template_text,
             op_summary_template_text,
+            may_snapshot_working_copy,
             may_update_working_copy,
             working_copy_shared_with_git,
         };
@@ -1158,6 +1180,8 @@ impl WorkspaceCommandHelper {
         } else {
             let hint = if self.env.command.global_args().ignore_working_copy {
                 "Don't use --ignore-working-copy."
+            } else if self.env.command.global_args().no_integrate_operation {
+                "Don't use --no-integrate-operation."
             } else {
                 "Don't use --at-op."
             };
@@ -1189,7 +1213,7 @@ impl WorkspaceCommandHelper {
         &mut self,
         ui: &Ui,
     ) -> Result<SnapshotStats, SnapshotWorkingCopyError> {
-        if !self.may_update_working_copy {
+        if !self.may_snapshot_working_copy {
             return Ok(SnapshotStats::default());
         }
 
@@ -1276,7 +1300,7 @@ impl WorkspaceCommandHelper {
         ui: &Ui,
         git_import_export_lock: &GitImportExportLock,
     ) -> Result<(), CommandError> {
-        assert!(self.may_update_working_copy);
+        assert!(self.may_snapshot_working_copy);
         let mut tx = self.start_transaction();
         jj_lib::git::import_head(tx.repo_mut()).await?;
         if !tx.repo().has_changes() {
@@ -1299,7 +1323,12 @@ impl WorkspaceCommandHelper {
             // state to it without updating working copy files.
             locked_ws.locked_wc().reset(&wc_commit).await?;
             tx.repo_mut().rebase_descendants().await?;
-            self.user_repo = ReadonlyUserRepo::new(tx.commit("import git head").await?);
+            self.user_repo = ReadonlyUserRepo::new(
+                self.env
+                    .command
+                    .maybe_commit_transaction(tx, "import git head")
+                    .await?,
+            );
             locked_ws
                 .finish(self.user_repo.repo.op_id().clone())
                 .await?;
@@ -2033,7 +2062,7 @@ to the current parents may contain changes from multiple commits.
             }
 
             #[cfg(feature = "git")]
-            if self.working_copy_shared_with_git {
+            if self.working_copy_shared_with_git && self.env.command.should_commit_transaction() {
                 let old_tree = wc_commit.tree();
                 let new_tree = commit.tree();
                 export_working_copy_changes_to_git(ui, mut_repo, &old_tree, &new_tree)
@@ -2041,8 +2070,10 @@ to the current parents may contain changes from multiple commits.
                     .map_err(snapshot_command_error)?;
             }
 
-            let repo = tx
-                .commit("snapshot working copy")
+            let repo = self
+                .env
+                .command
+                .maybe_commit_transaction(tx, "snapshot working copy")
                 .await
                 .map_err(snapshot_command_error)?;
             self.user_repo = ReadonlyUserRepo::new(repo);
@@ -2077,10 +2108,12 @@ to the current parents may contain changes from multiple commits.
             .map_err(snapshot_command_error)?;
         }
 
-        locked_ws
-            .finish(self.user_repo.repo.op_id().clone())
-            .await
-            .map_err(snapshot_command_error)?;
+        if self.env.command.should_commit_transaction() {
+            locked_ws
+                .finish(self.user_repo.repo.op_id().clone())
+                .await
+                .map_err(snapshot_command_error)?;
+        }
         Ok(stats)
     }
 
@@ -2222,7 +2255,7 @@ to the current parents may contain changes from multiple commits.
             .transpose()?;
 
         #[cfg(feature = "git")]
-        if self.working_copy_shared_with_git {
+        if self.working_copy_shared_with_git && self.env.command.should_commit_transaction() {
             use std::error::Error as _;
             if let Some(wc_commit) = &maybe_new_wc_commit {
                 // Export Git HEAD while holding the git-head lock to prevent races:
@@ -2244,7 +2277,12 @@ to the current parents may contain changes from multiple commits.
             crate::git_util::print_git_export_stats(ui, &stats)?;
         }
 
-        self.user_repo = ReadonlyUserRepo::new(tx.commit(description).await?);
+        self.user_repo = ReadonlyUserRepo::new(
+            self.env
+                .command
+                .maybe_commit_transaction(tx, description)
+                .await?,
+        );
 
         // Update working copy before reporting repo changes, so that
         // potential errors while reporting changes (broken pipe, etc)
@@ -2260,6 +2298,14 @@ to the current parents may contain changes from multiple commits.
         }
 
         self.report_repo_changes(ui, &old_repo).await?;
+
+        if !self.env.command.should_commit_transaction() {
+            writeln!(
+                ui.status(),
+                "Operation left uncommitted because --no-integrate-operation was requested: {}",
+                short_operation_hash(self.repo().op_id())
+            )?;
+        }
 
         let settings = self.settings();
         let missing_user_name = settings.user_name().is_empty();
@@ -3531,6 +3577,23 @@ pub struct GlobalArgs {
     /// implies `--ignore-working-copy`.
     #[arg(long, global = true)]
     pub ignore_working_copy: bool,
+
+    /// Run the command as usual but don't integrate any operations
+    ///
+    /// When this option is given, the operations will still be created as usual
+    /// but they will not be integrated to the operation log. The working copy
+    /// will also not be updated.
+    ///
+    /// The command will print the resulting operation ID. You can pass that to
+    /// e.g. `jj --at-op` to inspect the resulting repo state, or you can pass
+    /// it to `jj op restore` to restore the repo to that state. You can also
+    /// pass the ID to `jj op integrate` to integrate the operation.
+    ///
+    /// Note that this does *not* prevent side effects outside the repo. For
+    /// example, `jj git push --no-integrate-operation` will still perform the
+    /// push.
+    #[arg(long, global = true)]
+    pub no_integrate_operation: bool,
 
     /// Allow rewriting immutable commits
     ///