+++ cli/tests/test_completion.rs
@@ -228,6 +228,7 @@ fn test_bookmark_names() {
     --help	Print help (see more with '--help')
     --repository	Path to repository to operate on
     --ignore-working-copy	Don't snapshot the working copy, and don't update it
+    --no-integrate-operation	Run the command as usual but don't integrate any operations
     --ignore-immutable	Allow rewriting immutable commits
     --at-operation	Operation to load the repo at
     --debug	Enable debug logging
@@ -719,6 +720,7 @@ fn test_command_completion_short_name() {
     unset	Update a config file to unset the given option
     --repository	Path to repository to operate on
     --ignore-working-copy	Don't snapshot the working copy, and don't update it
+    --no-integrate-operation	Run the command as usual but don't integrate any operations
     --ignore-immutable	Allow rewriting immutable commits
     --at-operation	Operation to load the repo at
     --debug	Enable debug logging